requirements.txt
pandas
numpy
scikit-learn
yfinance
yfinance-cache
pycaret
feedparser
nltk
leia-br
tweepy
python-binance
